+++ R/auk-clean.r
@@ -56,7 +56,7 @@ auk_clean <- function(f_in, f_out, sep = "\t", remove_text = FALSE,
   if (!overwrite && file.exists(f_out)) {
     stop("Output file already exists, use overwrite = TRUE.")
   }
-  f_out <- normalizePath(f_out, mustWork = FALSE)
+  f_out <- normalizePath(f_out, winslash = "/", mustWork = FALSE)
 
   # determine number of columns
   # read header row
@@ -90,11 +90,10 @@ auk_clean <- function(f_in, f_out, sep = "\t", remove_text = FALSE,
                     list(sep = sep, ncols = ncols, print_cols = print_cols))
 
   # run command
-  err <- tempfile()
   exit_code <- system2(awk_path,
                        args = paste0("'", awk, "' ", f_in),
-                       stdout = f_out, stderr = err)
-  unlink(err)
+                       stdout = f_out, stderr = FALSE)
+  
   if (exit_code == 0) {
     f_out
   } else {